Allow tabbing to navigate past CodePane
CodePane in TutorialStep uses an editor to display the code. An editor
normally traps the focus since tab is a valid input command.

However for usage in CodePane, it's a readonly tool used to display code
with syntax highlighting. We want user to be able to tab to focus the
next component.

Add custom handling for this case.

Change-Id: I0fd1c0a221f0e84bb5263e47b8be85e881c61291
Fixes: b/67499214
Test: n/a