Begin separating tests into two groups
We want to separate tests into a sanity group
and a FAST group. Change Ieb640e3242a1b9e9b97aa14f602149e90d75fd32
configures the builder to run tests in both the old
and new groups, so tests will continue running as we
continue separating the tests.

As a test of this change, only
NewProjectTest.testCreateNewMobileProject is grouped under
the new sanity group.

Change-Id: Ifc5334e3dfdfae7351cdc2001cd7d83af55e21d1
Test: All existing tests