Fix GradleFiles from hashing non-existent files.
This deals with a bug introduced in ag/3128272
whereby files that do not exist on disk were
getting hashed causing intellij to represent
them as a black file. This resulted in a hash
being added when it shouldn't be.

In this change we make the checks on the file
more robust before we commit to hashing the
file.

Test: existing
Change-Id: I6395c74440714320374792c31204fda67700afd5