Android Studio can't find Add-ons when switching SDKs
A project with an add-on
(eg compileSdkVersion 'Google Inc.:Google APIs:24') will still not sync
when Android Studio switches from a SDK that does not have the Add-on
to another that has. To make it sync again, the compileSdkVersion needs
to be changed to the add-on target API (compileSdkVersion 24), sync, and
then back to the original value.

Bug: 69606910
Test: Run GradleSyncTest@sdkCreationForAddons() and added AndroidSdksTest@testCreateSdkForAddonCreatesItsParent
Change-Id: I0f7cb8f9175d05146727bc929db0a76a04edc272