Disable duplicateMessages test on Windows
BuildOutputParserTest#duplicateMessages simulates that some errors were
generated and pass them to BuildOutputParser. The output is done in a
way that the same error message will be generated from two different
parsers (MergingExceptionParser and BuildFailureParser) and the test
makes sure that only one is presented to the user.

The error messages contain the path of the files that generates the error,
but the path is processed and displayed in a way dependent of the OS. We could
make this test to be aware of what OS is it running but at the end there
is no difference in what the test covers on different OS. This CL
disables the test on Windows.

Bug: 69619617
Test: N/A disabling a test in Windows
Change-Id: I8d0080fc4fdb75cf750467587be18986380e7761