Moved HttpData and related classes into a subpackage
The number of classes associated with HttpData is growing (and will
get a new addition in a followup CL). Now seems like a good time to
create a subpackage for HttpData.

Despite all the noise in this CL, there is no new code added
(except a few methods made @VisibleForTesting public instead of
package private, due to the move). Everything else is just a
straight refactoring.

Test: Covered by existing tests.
Change-Id: I8f2c6c3bc29d3b39ee114db105fc6e05588d42b2