No longer allow user to press "Next" before step is ready
If you currently spam on the "Next" button in a wizard
dialog, you can actually press it (if you're fast enough,
or just unlucky) before the next step finishes propogating
its "you can't go forward" state to the UI.

In other words, the wizard's state, itself, would be up
to date, while the UI itself wouldn't get notified until
after an enqueued event was fired.

To resolve this, we just get a little less strict with the
logic - return a boolean instead of throwing an exception.
To the user, what this means is, they can *potentially*
press the next button for a frame or two where it doesn't
do anything. This would be the same if we disabled the
button for a frame or two. But because the time scale
is so short, it's unlikely anyone will ever notice.

Test: Updated
Fixes: 68972016
Change-Id: I3ed1ae3b49e86ecedbb1dcd80edc99058ace4d1b