Remove NlModel from SceneView
This CL is for decoupling the dependency between
NlModel and SceneView.

DesignSurface may have multiple SceneManagers and
NlModels in the future, thus we must put SceneManager
in the constructor of SceneView to indicate which
SceneManager it belongs with.

Also, SceneView is designed for displaying View in
DesignSurface. It should not have information of
NlModel. For back compatibility, delegate the
SceneView.getModel() to SceneManager.getModel() and
mark it as deprecated.

Test: Updated related tests.
      Passed all tests in NavigationTestCase.
      Paased all tests in LayoutTestCase.
Change-Id: Ic0bd86ec7587cbd42ecdb74f3f33fc1c60279026