Refactor MethodModel and related classes
MethodModel started as our representation of a Java Method. However,
adding simpleperf (and now we're adding systrace) as a profiling option
caused this class (and a lot of other classes that depend on it) to be
modified in order to support characteristics of other things, such as
native functions and syscalls.

This CL splits MethodModel into JavaMethodModel, NativeFunctionModel,
SingleNameModel and DummyModel, and update their usage across profilers.
Splitting MethodModel is also important because we keep adding fields to
it that might be specific of only one type of data (e.g. .so files for
native fuctions) that can be represented by the model.

This is the first step of the refactoring, but it's not the last one. We
still need to rename MethodModel (maybe to CaptureNodeModel?), as it gives
the impression of representing only Java methods, which is not true.
Next, we should revisit CodeLocation/CodeNavigation to decide if we need
differenti classes for native and java code navigation. Therefore, the
MethodModel refactoring bug shouldn't be closed yet when this CL gets
merged.

Bug: 69904551
Test: Update existing tests
Change-Id: I6627fd4e1fe0e80c752d179c1b66fb35e9c93e1b