[New PSD] UI tests base and an example.
The CL includes some base fixtures and utility classes for the UI tests of
the New PSD as well as an example test which ensures that the module list
selector is well-behaved.

The CL also sets component names for some UI containers which are difficult
to identify in the tests otherwise.

Bug: 69788672
Test: BasePerspectiveConfigurableFixture
Change-Id: Id94b0cdd801dac74ab0813409973e6c5d354d7b0