Clean up around repo for the gradle plugin.
We had several versions of the in different places.
Move them all to //tools/base/build-system, and make
it clear what they contain.

Also update all rules and code to deal with the removal
of the automated _repo suffix added to all repo.
It now only add the .zip extension.

Change-Id: Icfd15ad2ee3899489e70a1bbdc1b86bcebbfd6fa
Test: existing tests.