Apres-merge fixes for IndexingSuspender service.
This CL achieves the following:

(1) Enters dumb mode only on project setup, which allows
droppping the complicated logic dependent on project being
initialised first - this is actually what broke with the merge;

(2) Refactor the activation conditions into one method, and
look at them in the constructor as well, in order to avoid
subscribing to events when we don't need them;

(3) Refactor the way the instance is initialised in order to accommodate
the surfaced fact that the service instance that we 'replaced' with
IdeComponents in the integration test actually remained
alive (expectedly), but also remained subscribed to all the
activation/deactivation events (unexpectedly), so that in test we
ended up with two active instances of the suspender service.
This wasn't the case in IJ2017.2 for some reason or it wasn't visible
because the state machine transitions were agnostic to that.

Test: Existing. Adjustments in the service initialisation code made
IndexingSuspenderTest pass again for the new suspension logic,
without changing the test code itself.
Change-Id: Ie20f8e3b549130024e73ae0997286c86acbf3142