Add recoverWhile rules to subqueries and WITH clause
This way anything between a 'WITH' and 'SELECT' is ignored if is not
valid, but the second half of the query gets parsed so codecompletion
etc. works as expected. The same applies to subqueries: everything
inside the parenthesis will be consumed even if it's invalid.

GrammarKit error reporting gets confused by backtracking over a rule
with recoverWhile, so I extracted the WITH clause out of the rules that
use them, so we don't reparse it over and over again.

Test: RoomSqlParserTest
Change-Id: I620a59f8ba065123f8dcb67ff6361dd39e899711