Replace DelayedInitialization with ClearResourceCacheAfterFirstBuild
DelayedInitialization is responsible for clearing the
resource cache after source generation is complete and for waiting
to initialize the designer and preview
tabs of the layout editor until after the resource cache has been
cleared. Because its implementation is heavily dependent on Gradle,
DelayedInitialization is currently disabled for Blaze projects.

This CL resolves the issue by introducing
addSourceGenerationCallback(), a build system-independent
API in ProjectSystemSyncManager that allows callers to specify a
callback to run once the first project sync with source generation
is complete.

This CL also replaces DelayedInitialization with
ClearResourceCacheAfterFirstBuild, which uses
addSourceGenerationCallback() to clear the resource cache at the
appropriate time (if necessary) and provides an API for executing
callbacks after the cache has been validated.

Bug:  65213155
Test: GradleProjectSystemSyncManagerTest
      + ResourceCacheClearingComponentTest

Change-Id: Ibca2cac11933f6daa91eab17ca167a1d6d6e53dd