Refactor MethodModelHRenderer to render models differently
We have different types of CaptureNodeModel and they are rendered
differently in HTreeCharts. MethodModelHRenderer shouldn't be responsible
for handling all types of model. Instead, it should identify the model
type being rendered and call a corresponding class to retrieve the border
and fill colors of each rectangle corresponding to a node model.

This CL also splits the Native models into C/C++ function model,
Syscalls model and Non-symbol models. The idea is to be able to
represent these models differently later, maybe with a different color
in the HTreeChart, or with a label in its tooltip distinguishing them.
For now, the Chart functionalities and colors remain intact.

Bug: 69864355
Test: Made MethodModelHRenderer testable and added the tests for it.
      Also added unit tests for new classes. Finally, updated existing
      tests, such as NodeNameParserTest#testCppMethodsParsing.

Change-Id: I6b22571652ce307f2e87cac86a1c3ee3405c2f71