Remove references to groovy classes from JavaVersionDslElement.
This change removes one of the last two references to groovy specific
classes from the GradleDslElement subclasses.

This involved adding extra information to the GradleDslElement,
a use assignment field. This is a boolean that represents whether
the element should be written to the file using the assignment syntax
(e.g "prop1 = 'value1'") or the application syntax (e.g "prop1 value1").
The value of this field is set by the parser, if the element was
parsed from an application statement and can be changed via
GradleDslElement#setUseAssignment().

Bug: 67846255
Test: Existing
Change-Id: I6aa84dec0ed1c30fe3fcd4f664b919c88ce8bbb6