Do not wrap checked icon with PoppedIcon in designer.
The IJ ActionMenuItem would wrap the selected icon from a
ToggleAction in a PoppedIcon which raises the icon to make it
look like a button.
This works great for a toolbar button but not for a menuitem in
a popup menu.

Changed the code in NlActionManager to not use a ToggleAction to
avoid the raised icon in the popup menu.

Test: NA. Visual check for now.

Change-Id: Ia545df45ebf0577c09937c86b0d0f896d275d3ae