Use SmartPsiElementPointer inside of NlComponent
The XmlTag stored inside NlComponent can become invalid, which can then
trigger IDE errors. So instead use SmartPsiElementPointer which will
always reference a valid version of the XmlTag, assuming it still
exists.
Because of the wide usage of NlComponent.getTag and the fact that it is
annotated @NotNull, we have to be able to return something whenever the
SmartPsiElementPointer.getElement would return null. That is why we keep
for the moment the reference to the original XmlTag. That is a temporary
fix as a more complete fix would be too risky for now.

Bug: 70264883
Bug: 70300930
Test: designer tests, ComponentTreeTest, NlEditorTest
Change-Id: I173a4024d2a5c390842320bd131ed05558530abe
(cherry picked from commit 8e42040d015e2e0bc25ff5b5be7902478c684e57)