Assertion failed: "Don't register startup activities for the default project"
When Studio is built, it runs the IDE in a special mode in order to
iterate through the options panel and build up a search index.

This was throwing
ERROR: Assertion failed: Please don't register startup activities for the default project: they won't ever be run
java.lang.Throwable: Assertion failed: Please don't register startup activities for the default project: they won't ever be run
    at com.intellij.openapi.diagnostic.Logger.assertTrue(Logger.java:174)
    at com.intellij.ide.startup.impl.StartupManagerImpl.checkNonDefaultProject(StartupManagerImpl.java:80)
    at com.intellij.ide.startup.impl.StartupManagerImpl.registerPostStartupActivity(StartupManagerImpl.java:99)
    at com.android.tools.idea.gradle.project.sync.GradleFiles.<init>(GradleFiles.java:84)
because it's being invoked on the default project.

This CL makes the GradleFiles handler skip the default project.

Test: Unit test included
Change-Id: I178b13d33470d8a71999edb4dc8835736e86a139