Make sure all documents affected by the build model changes are saved.
This change constructs a list of all the PsiFiles that could be involed
with modifications to the given build model. This are then all unblocked
by completing any outstanding psi opreations and saved to disk using
FileDocumentManager#saveDocument.

This change also enables
GradleSyncIntegrationTest#SyncWithAARDependencyAddsSources which was
failing as a result of the parents build file not being saved.

Bug: 70205737
Test: Existing
Change-Id: I5c7e9c0f9faaf1c3dcc6c65d5c5ae4ffb73a1408