Delegate getData to DesignSurface's LayeredPane
The refactor actions were disabled in the component tree because the
data context was coming from the tree and needed to come from the
LayeredPane to be correctly populated.

Also since the getData method was the same in ComponentTree and
DesignSurface, the tree now just delegates the getData calls to its design
Surface.

Bug: 68712729
Test: N/A
Change-Id: I3143de680f9d53b205e77be79b999096c297911b