Fix Kotlin instrumented tests run by context
When creating android test run configurations are created by context for
kotlin code, the corresponding parent PsiElement was not always found.
This CL fixes it by searching for the parent of the corresponding
UElement or the corresponding UElement of the parent.

As a secondary problem, Android Test configurations were not always
replacing old Unit test ones. This CL also fixes this problem.

Test: AndroidTestConfigurationProducerTest and
AndroidJUnitConfigurationProducersTest
Bug 66084314

Change-Id: I46704d22932bed191eb5aba260b01f0c7e964ed6