Make IdeComponents#mockProjectService undoable
IdeComponents#mockProjectService passes null for the undoQueue
argument of doReplaceService(), which means that restore() does
not undo the mock. This causes problems when we need one test to
mock a project service while another test in the same class relies
on the actual implementation of that service.

This CL makes it so that mockProjectService() passes myUndoQueue
to doReplaceService() instead of null so that restore() undoes
mocks added by this method.

Test: affected tests still pass
Change-Id: I203bce4872c0847ef77a44197d145b4b6ae1c915