Fix race condition between project initialisation and batch update start.
There were some cases reported when IDE error log would be populated
with assertions related to incorrect batch update depth. After further
analysis, the issue appeared to be down to a race condition between
project initialisation and batch update messaging, because when
a notification about batch update start was sent, it was possible
that the module root manager was not initialised yet, but later, when
a batch update finish event was broadcast, the module root manager
would already receive it and on attempting to decrement the batch
update depth, which would unexpectedly be zero at that point,
because batch update was never entered.

This CL brings back some project initialisation control logic
and improves it to make sure we activate indexing suspender not
earlier than project is initialised.

Test: Existing.
Change-Id: I558a38bec5026276576b9c8255a9ee259ddb0617