Close "Parsing capture..." panel if parsing fails
If an exception was thrown during a capture was being parsed, we would
just add a warning to the logs, but keep the panel displaying the
"Parsing capture..." animation open. We need to make sure to close the
panel, so users don't think it is still being parsed.

This CL adds an intermediate capture state (PARSING_FAILURE) and fires
the CAPTURE_STATE aspect when this state is set. Also, we removed the
call to updateCaptureState from CpuProfilerStageView#updateCaptureSelection
because the former should be triggered by CAPTURE_STATE aspect, while
the latter should be triggered by CAPTURE_SELECTION.

Bug: 70274021
Test: CpuProfilerStageTest#parsingFailureIsNotifiedToUi
Change-Id: I429b0451833719fe343107b9fefb7dde2ebfdcf6