Fix API contract violations in AndroidProjectStub
TestProjects.createFlavorsProject was creating a project in which
a single variant contained multiple flavors, which would require
the flavors to belong to separate dimensions. However the project
was incorrectly reporting those flavors as belonging to the same
dimension (or rather, it didn't create any dimensions or do any
associations between dimensions and flavors, which is how legacy
models reported flavors belonging to a single dimension). This
patch updates the model to include two flavor dimensions and
associate each flavor with the correct dimension.

The AndroidProjectStub.forEachVariant method is now functional.

ProductFlavorStub now has the capability of reporting an association
with a flavor dimension.

VariantStub will now report an android test artifact, if one exists.

Test: This whole patch is test code.
Change-Id: I699a880a76ce28e9339cc509c9a8d0e18e350605