New Project windows are off when changing the IDE font size
"Some devices require additional SDKs..." is not wrapping because
Swing labels text needs to be <html> for wrapping to work.
"By targeting API ... " is <html> but the panel layout (FlowLayout) is
not forcing the component bounds and the width just keeps growing.

Test: N/A No code changed, only UI layout was updated, tested in
normal/presentation mode, and with a big IDE font
Fixes: 68339663

Change-Id: Ie93fb4a0f3f4e700328acd0605d7534584493a9b