Remove the timer from TranslationsEditorTextField
Because it causes write unsafe context exceptions to be thrown

We didn't want to synchronize the value in the text field at the bottom
of the translations editor with the table after every keystroke because
that used to rebuild the data model every time. The timer was put in
place to trigger a rebuild when 500 milliseconds have elapsed between
keystrokes.

I removed the ResourceChangeListener that triggered the rebuilds in
a7a19ff126 so we don't need the timer anymore

Bug: http://b/70295823

Test: TranslationsEditorTest
  .enteringTextInTranslationTextFieldUpdatesTableCell

Change-Id: I8466a50c6adef36ede6a07d4132777f6c8a25a99