Make project-system-gradle sync APIs wait for source generation
A common reason for programmatically requesting a sync is that a
dependency has been automatically added for the user (e.g. after
adding a component from the layout editor palette with a missing
dependency). In such cases, callers usually want to attach a
callback which will execute after both the requested sync and the
subsequent source generation complete. The project system API
currently provides no way to wait for source generation to complete.

This CL resolves the issue by updating GradleProjectSystemSyncManager
so that it does not consider a sync with source generation to be
complete until after the source generation finishes. This means that
callers who need to wait for source generation can use the future
returned by syncProject().

Because the project system structs are not initialized until the
first time the project system APIs are used, we can not handle waiting
for source generation before broadcasting to PROJECT_SYSTEM_SYNC_TOPIC
inside GradleProjectSystemSyncManager, or else we would miss any sync
which takes place before the sync manager is initialized. Instead,
this CL introduces GradleSyncResultPublisher, a project
service residing in the gradle package which is responsible for
listening to build and sync results and broadcasting to
PROJECT_SYSTEM_SYNC_TOPIC appropriately.

GradleSyncResultPublisher is initialized by GradleSyncState once the
first Gradle sync has started. This ensures that we don't attach a
listener to dead message bus topics for non-Gradle projects.

Test: Updates to GradleProjectSystemSyncManagerTest
      + existing tests cover Gradle changes
      + verified that GradleSyncResultPublisher
        is not initialized for Blaze projects
Change-Id: Ie1f44a7a2433f688bd7f0b4001af8e31b6dfe81c