New Activity Wizard Fails with custom named gradle build file
Gradle modules by default have a "build.gradle" file on the module root.
The name and location of the build file can be modified by changing
the project settings (settings.gradle), for ex adding:
   project(':mylibrary').buildFileName = 'mylibrary.gradle'
Adding a new Activity to the library with the renamed build file was
adding its dependencies to the wrong file ("build.gradle" instead of
"mylibrary.gradle").
This CL makes sure that the Module information is made available when
adding dependencies, so we can find the right build file.

Test: Added AndroidLibraryTest@testRenameLibraryAndAddActivity
Change-Id: I758340a07c5c2f1a9909f5cc3ecbb8613811a68b
Fixes: 68150753