UI Tests: Fix NewActivityTest@createActivityWithNonDefaultPackage
Adding a new Activity runs a new Dialog and the "Package" text field is
implemented by an IJ EditorComponent.
The IJ EditorComponent executes setText() on a transaction Guard, and
the real setting of the text is postponed until the dialog is dismissed.
Calling textField.setText() on a TransactionGuard.performUserActivity()
simulates an user action, that ignores the ModalityState, and sets the
text as expected.

Test: Run NewActivityTest and other tests that call replaceText()
Change-Id: Id42bf17ff3637fdac6968fd69924afd5f70a6598