Add a new hyperlink when project is not initilized and using older plugin.
Current quickfix adds an hyperlink to the file containg the outdated
plugin version, but this information can only be retrieved if the
project is initialized and because of that, ag/3363258 causes the result
to be different on a freshly imported project.

This CL checks adds a new hyperlink that solves what file to open only
when the user clicks the hyperlink. From a user perspective, there is no
difference betweeen having or not an initilized project.

Also added a test to confirm the correct hyperlink is created depending
on project initialization.

Test: New test.
Bug: 70568970

Change-Id: If9e63085eb157d9e39437507e78289c63b87e835