Preserve selection in Android view during Gradle sync
During Gradle sync AndroidModuleNode or a NdkModuleNode could become
a NonAndroidModuleNode and vice versa. To preserve selection in
the Project panel, nodes of these three types should be considered equal
if they represent the same module. This CL establishes equality
relationship between instances of AndroidModuleNode,
NonAndroidModuleNode and NdkModuleNode representing to the same module.

Test: AndroidViewNodeEqualityTest
Bug: 70635980
Change-Id: Ia9281a0269c5ecebbc0e6338fd6928b36ac37001