Support replacing dependencies by PsiElements.
This change extends the DependenciesModel to allow
replacing of dependencies by the use of their PsiElement.
This enables it to be used in conjunction with intellijs
apis.

Also, we add a getPsiElement method to the DependencyModel
to for the same reason.

We have made sure this works with the following dependency syntaxes:
  1) Application statements, with a single or multiple args.
  2) Method calls, with single or multiple args plus closures.
  3) Both map and compact notations.

Bug: 70561159
Test: Unit tests added
Change-Id: I74abc977e41e97127b1feba0a15f4cdcc7c90c63