Fix createCppKotlinProject test
bug: 70339267

As observed locally, after it clicking on "Create Virtual Device..." in
Avd Edit Wizard dialog, next step dialog is not showing up.
Adding the step that closing the cpp file (the "editor" doesn't have
the focus any more) before creating a new AVD
prevent that happens. It might because the "editor" has the focus after
opening the cpp file.
As test it manually or run this test in debug mode with breakpoints, no
such issue.

Test: ran this test, passed, will als monitor it on build bot
Change-Id: Ia03a6ef5373174cdce63e0cff8de72b708bde0d3