Ensure correct property types are set.
This change ensures that correct property types are
obtainable from the GradlePropertyModel.

To do this we move the PropertyType enum into a
new class, change the holder from a GradlePropertyModel
to a GradleDslElement and set this element type when
parsing the build file instead of on creation of the
model.

Bug: 67519210
Test: Corrected existing unit tests.
Change-Id: I3ce4ab717b31dfda47c5c5e41e547f3d15bec862