Hitting a component if a parent is selected
Right now in the nav editor, the global action arrow isn't considered
"hit" if you click on it or hover over it and:
- it's over another destination
- it's usually drawn underneath that destination (that is, the
destination associated with the action appears before the other
destination in the file)
- the destination associated with the action is selected, brining it to
the front.

The arrow itself is drawn on top, but since (currently) the arrow is
drawn outside the bounds of the parent component, the parent isn't
marked as hit, and therefore the existing logic doesn't cause the arrow
to be selected as the best choice.

Test: pending approval of the logic
Change-Id: If01a08e0e286bb6063608eca16ba8e9bdcf0ed90