Temporary workaround for font download problem.
When we are downloading fonts we are experiencing IOExceptions in:
   NetUtils.copyStreamContent
where the compressed stream length doesn't match the content length
in the header of the response. The error is something like:
   "Connection closed at byte 3548. Expected 3551 bytes."

This temporary fix is to try to read the font without compression
if we get an error reading with compression.

Alternatively we could omit the lines in NetUtils that checks the
content length - but that seems wrong to me.

Bug: b/69217300

Test: NA, the issue has not been reproduced locally.
Change-Id: I12bb9114757b0140b898a852fdd936442d01929a