Set sync result future using PROJECT_SYSTEM_SYNC_TOPIC
Currently, GradleProjectSystemSyncManager#syncProject uses the
GradleSyncListener parameter of GradleSyncInvoker#requestProjectSync
to set the value of the sync result future after the sync has
completed. As it turns out, the listener parameter does not receive
notifications that are consistent with those broadcast to
GradleSyncState subscribers (e.g. if the sync cannot resolve a
dependency, the parameter will be told the sync was successful, whereas
subscribers will be notified of the failure), and will likely be
removed in the future (see comments in ag/3379571).

This CL ensures that the sync results produced by the project system
API are consistent by changing the future-setting logic in
requestSync() so that it relies on PROJECT_SYSTEM_SYNC_TOPIC instead
of the listener parameter to GradleSyncInvoker#requestProjectSync.

Test: GradleProjectSystemManagerTest still passes.
Change-Id: I109f45d431712d826baec84224883fb69961042c