Change FlavorType models to use GradlePropertyModel.
This change alters the FlavorType, BuildType and ProductFlavor
model to return ResolvedPropertyModels instead of GradleValues. This
allows us to expose more information about the properties.

We create a ResolvedPropertyModel subclass of GradlePropertyModel in
order to allow users of the models to get the actual value in one call
to #getValue. The standard GradlePropertyModelImpl does not allow this
as it saves the type as a REFERENCE and its #getValue will be the
reference text.

We also take this opportunity to clean up the models interfaces by
removing any set* methods (since they can now be set through
GradlePropertyModel) and removing the ability to chain operations.

In order to reduce confusion GradlePropertyModel#getUnresolvedValue has
been renamed to GradlePropertyModel#getRawValue

Bug: 71573751
Test: Changed existing tests
Change-Id: Ic23b58b6c3bdc3a35c7c0c06e86b89bb3747bf76