Wait for the application to launch on emulator
LayoutInspectorTest occasionally activates the Layout Inspector before
the app has launched on the emulator. This causes the test to fail.

This change makes LayoutInspectorTest to wait for the app to launch on
the emulator before continuing. The wait is implemented using
uiautomator. The wait objective is satisfied when uiautomator returns
information about views that belong to the test application.

Bug: 71361448
Test: LayoutInspectorTest
Change-Id: I50612263e690b5ca04a6eda0ff0d7b184c8262c6