[AVD Mgr] Show the correct device type with each AVD
In the list of AVDs, we show an icon with each AVD. The icon
indicates if the AVD is a phone/tablet, TV or Wear device. These
icons were cached incorrectly, so all devices were given the same
icon.

This CL corrects the icon caching. Now each device gets the icon
it deserves.

Bug: 71493172

Test: Added AvdDisplayListTest.testGetDeviceIconPair

Change-Id: Ib70d0146794ddb506589089ac1444514dbc5da53