Creates an EmptyPropertyModel class.
This class represents a non-existent property inside the
GradleBuildModel, this will be used by other parts of the model to
indicate that a property doesn't exist. It allows the user of the
model to set a value if they wish.

A new GradlePropertyModel.ValueType of "NONE" has been added to by
used by these empty models.

Bug: 67519210
Test: Unit tests added
Change-Id: I4d43f70dac2ce6de2fde9f59f313800bb9debc0d