Don't overflow the stack when a cycle exists.
This changes stops the IDE from throwing a stackoverflow when the Gradle
build files contained a cycle of property dependencies. We add a check
that looks for any cycles before attempting to resolve a value.

To support this we have made GradleDslLiteral#getExpression return the
currently set PsiElement rather than the last commited one. In order to
allow the GroovyDslWriter to query for both,
GradleDslLiteral#getLastCommittedValue was added that only returns the
PsiElement that is present in the file.

Bug: 70335549
Test: Unit tests added.
Change-Id: Ia5148ce01a582471964b889c1196c7aea0d545d2