Rename ModulesByGradlePath to ModuleFinder.
ModulesByGradlePath maintains a map from Gradle path
to module.
But with composite build, Gradle path can not uniquely
identify a module, because two modules in different
projects can have the same Gradle path.

This class needs to be enriched to find modules from
included builds as well.
Current class name does not reflect the actual mapping,
rename to ModuleFinder, and keep methods like
findModuleBy***.

This CL only does the renaming, the actual support of
composite builds will be in the next CL.

Bug: 69638926
Test: current tests.
Change-Id: I3a5bacd07af7c0ce29ec525b123916e523488c83