Make ModuleResourceManagers to be a ModuleService
AndroidFacet has an unnecessary dependency on ModuleResourceManagers.
The only reason for that is that some cached data (resource managers)
inside that class is explicitly cleared by the facet when the platform
or source roots change.

This CL converts the ModuleResourceManagers from an
AndroidFacetScopedService to just a ModuleService, and removes the
dependency from the facet to the resource managers. The
ModuleResourceManagers directly listen for root changes and perform
their internal cleanup as needed.

Bug: 71553467
Test: Existing tests, no new functionality

Change-Id: If90eea6b448a2e7844e2e2fc9029a6abd5f2bc3e