Make AndroidFacet independent of "ModuleSourceAutogenerating"
Android projects that use the IntelliJ build system require the IDE to
automatically generate sources corresponding to AIDL, RS, etc files.
ModuleSourceAutogenerating is part of the infrastructure that supports
this source auto generation.

The primary purpose of this CL is to remove the dependency from the
AndroidFacet to this ModuleSourceAutogenerating class. The dependency
was present for two reasons:
 1. Initialize the source generator during initialization of the facet.
 2. Reset the source generator state when the project roots change.

Both of those are eliminated in this CL: The first dependency is removed
by having the initialization happen during object construction, and the
second one by having the source generator itself listen for project root
changes.

Bug: 71553467
Test: GradleSyncIntegrationTest has been updated, manually confirmed
that IntelliJ projects still have the source generator turned on
Change-Id: Ia61ac51868f6afbca086dde7ad1b1321180251cb