[NPW] new Module incorrectly generates package in AndroidManifest
This CL removes a memory optimization done for ChooseModuleTypeStep. The
same NewModuleModel object was shared by all "New Module" steps, and
that causes side effects when updating one UI step leads to update the
"model" data that belongs to other steps.
In the example of this bug, there are a few Steps with a similar UI
(Phone/Table Module, Feature Module, Android Wear Module, ...). This
steps "conditionally" bind the "module name" with the "package name".
The "conditional" field used to bind these two model fields is owned and
created by each individual step. When the user manually changes the
value of "package name", only the active step breaks the "conditional"
binding. Then if the user changes "Module name", the model
"package name" is still updated, by the remaining Steps where the
"conditionally" bind is still allowed.

Test: Run NewModuleTest, other UI tests should keep passing
Bug: 70989920
Change-Id: I8fa04650a96c9e89cc6e4e9d6f10500e308f65bd