Make the slider in ReferenceEditor a smaller height on Mac.
Without this change the slider in the ReferenceEditor could cause
a scrollbar for the inspector to show, which in turn may hide the
slider and cause the scrollbar to hide (In a loop).

This fix will make the slider smaller such that it will not
trigger the need for a scrollbar on the inspector.

Bug: b/71491838
Test: Manual for now.
Change-Id: Iaf4cf6187d6e56394b4de9241dca7d7cac4dfaae