EspressoRecorderTest refuses to be stable
Repeatedly clicking the editor to move the text caret inside the
text editor does not resolve the flakiness of this test.
Screenshots from previous test runs suggest that the text caret
does end up where it needs to be, but the test still fails.
Perhaps the position of the caret that we check becomes out of
date by the time the editor finishes loading or modifying the file.
We need a fixture method that waits for the IDE to stop changing
the file. Until then, we should use a workaround for the
EspressoRecorderTest

Test: EspressoRecorderTest
Bug: 37150124
Change-Id: I8003a3b35234b4e691a1412f12ddabe3d29fb9d6