Fix ApkViewerTest and others
Several UI tests have been having trouble navigating to a path in the
project view, due to a bug in the implementation of
ProjectViewFixture.waitForTreeToFinishLoading(). This previously tested
for the absence of the spinning wheel; however, this icon is not
guaranteed to be displayed, especially for short loading tasks. This
resulted in the waitForTreeToFinishLoading() method to return
immediately, even though the tree was not done loading.

This CL replaces that behavior with waiting for the JTree's model's
isProcessing() method to return false.

Bug: b/70580328, b/70633876
Test: ApkViewerTest and others
Change-Id: I28df4638a3bbfbcc70442087e2d3af25e6f15af2