Change ModuleDataBinding to be a ModuleService
Currently, it is an AndroidFacetScopedService, but in reality, there is
no state in this class that is specific to the AndroidFacet. This CL
simplifies the implementation quite a bit, and the new implementation
makes it obvious that this is purely a container that stores whether
databinding is enabled or not, and a reference to the LightBrClass per
module.

Test: existing tests updated, one test removed because it wasn't useful
Change-Id: I52cb972dadf6653e460fee0ca5967ba43c5f5b8f