Try one more time to click Next button
NewProjectTest.changeLibraryModuleSettings failed sometimes because the
dialog for "Target Android Devices" step is not showing during "Create
New Project".
Here is a workaround: Try more times to click Next button to wait for "Target
Android Devices" step to show.

Change the timeout in getConfigureFormFactorStep() to default: 10
seconds.

Test: ran NewProjectTest.changeLibraryModuleSettings test, will monitor
it on build bot

Change-Id: I8a3a121fef691f28f97a317e10889e6fa4ee0a45