Don't have InstantAppRunTest move too early
InstantAppRunTest.createAndRunInstantApp can sometimes click on
UI buttons before they are are ready. This is usually caused by
the test performing actions before the IDE is ready for such actions.

In this particular bug, the run button is disabled until the IDE
is sometimes not done with background work. This causes the test
to be flaky.

The solution is to make the test wait for background work to finish
before continuing.

Bug: 70567643
Test: InstantAppRunTest.createAndRunInstantApp
Change-Id: Ic421d68c4234a0d25c985aff5758364537e4e213