Avoid logging exception if device is not online
Since device properties can be null (e.g. device is offline), we should
not log an error as it is expected behavior.

Instead of logging an error, we simply cache the value and return
"null" values to the caller, as callers are already ready to deal
with "null" values.

Ask taskExecutor to start retrieving task everytime but only update when
values changes (e.g. from offline to online). This will make taskExecutor
busier but help to avoid async problem

Test: Updated existing text to reflect behavior
Bug: 71766662
Change-Id: I6b696e2eafce375ca0a46ca6e959b4cca772abfa