UX: Show flag for locale depends on the setting
Change the locale menu style in Layout Editor depends
on the setting of flag appearances.

The setting can be found in "Setting -> Appearance & Behavior
-> Appearance -> Show Flags for Languages"

When displaying flags is enable, mark the current locale and
display the flag for other locales except the default locale.

When displaying flags is disable, mark the current locale only.

Test: manually test.
Bug: 69427340
Change-Id: I1124b1c105e08cd433ff5814ca8e3a2f5f14c918