Merge EmptyPropertyModel into GradlePropertyModelImpl.
This change merges the EmptyPropertyModel into GradlePropertyModelImpl,
this allows us to be more flexible when converting between properties
that exist on file and properties that don't. This enables us
to present a much clearer interface to callers.

We also fix the issues with deleting properties after this have been
set, as well as issues when deleting propertyModels that are derived
from lists.

Bug: 67519210
Test: Most are existing added one more to test new logic
Change-Id: I40541eed92ea16b80c2c53bca15adabb65703fb7